courtney-free/api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pandas==2.2.2
numpy==2.0.1
httpx==0.27.0
email-validator==2.2.0
aiosmtplib==2.0.2
python-multipart==0.0.9